refactor(controller) : LottoController,ProfitController로 세분화
