+++ src/main/java/lotto/controller/LottoController.java
@@ -1,16 +1,11 @@
 package lotto.controller;
 
-import static lotto.util.Constants.ZERO;
 
 import camp.nextstep.edu.missionutils.Console;
 import java.util.List;
-import java.util.Map;
-import java.util.Optional;
 import lotto.model.BundleGenerator;
 import lotto.model.LottoBundle;
 import lotto.model.LottoDTO;
-import lotto.model.ProfitCalculator;
-import lotto.util.enums.LottoResult;
 import lotto.util.exception.parental.InputValidationException;
 import lotto.util.exception.parental.MathOperationException;
 import lotto.model.BonusNumber;
@@ -22,7 +17,6 @@ import lotto.view.OutputView;
 public class LottoController {
     private static final boolean TRUE = true;
     private static final boolean FALSE = false;
-    private static final Long LONG_ZERO = 0L;
 
     boolean validInput = FALSE;
 
@@ -33,15 +27,17 @@ public class LottoController {
     PurchaseAmount amount;
 
     public void winningResult() {
-        // 몇 개 구매 했는지
+        /*로또 구매 금액 입력, 검증, 로또 번호 생성*/
         purchaseLottoTickets();
-        // 구매 개수에 맞게 로또 번호 출력
+        /*로또 번들 생성, 로또 번호 출력*/
         generateLottoBundle();
-        // 정답, 보너스 번호 입력
+        /*당첨 번호 입력, 검증*/
         setWinningNumbers();
+        /*보너스 번호 입력, 검증*/
         setBonusNumber();
-        // 당첨 통계 출력
-        displayProfitReport();
+        /* 당첨 통계 출력*/
+        ProfitController profitController = new ProfitController(lottoBundle, amount);
+        profitController.displayProfitReport(winningNumbers, bonusNumber);
     }
 
     private void purchaseLottoTickets() {
@@ -91,46 +87,5 @@ public class LottoController {
                 System.out.println(e.getMessage());
             }
         }
-    }
-
-    private Map<String, Integer> displayStatisticReport() {
-        OutputView.winningStatistics();
-        OutputView.separator();
-        Map<String, Integer> compareResult = lottoComparator();
-        compareResult.forEach(OutputView::getResult);
-        return compareResult;
-    }
-
-    private Map<String, Integer> lottoComparator() {
-        return lottoBundle.compareLotto(winningNumbers, bonusNumber);
-    }
-
-    private void displayProfitReport() {
-        Map<String, Integer> compareResult = displayStatisticReport();
-        long totalPrize = calculateTotalPrize(compareResult);
-        double profit = calculateProfit(totalPrize);
-        OutputView.getProfit(profit);
-    }
-
-    private long calculateTotalPrize(Map<String, Integer> compareResult) {
-        return compareResult.entrySet().stream()
-                .filter(this::shouldIncludeEntry)
-                .mapToLong(this::calculatePrizeForEntry)
-                .sum();
-    }
-
-    private boolean shouldIncludeEntry(Map.Entry<String, Integer> entry) {
-        Optional<LottoResult> result = LottoResult.fromDescription(entry.getKey());
-        return result.isPresent() && entry.getValue() > ZERO;
-    }
-
-    private long calculatePrizeForEntry(Map.Entry<String, Integer> entry) {
-        Optional<LottoResult> result = LottoResult.fromDescription(entry.getKey());
-        return result.map(lottoResult -> (long) lottoResult.getPrize() * entry.getValue()).orElse(LONG_ZERO);
-    }
-
-    private double calculateProfit(long totalPrize) {
-        ProfitCalculator profitCalculator = new ProfitCalculator();
-        return profitCalculator.calculateProfit(totalPrize, amount);
     }
 }
